+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   
   * New Features:
     * `faspex5`: added command `admin configuration` for global parameters.
+    * `node`: "gen3" browse now returns all elements using pages, and supports option `query` with parameter `recursive`, `max`, `self`.
   * Issues Fixed:
     * `faspex5`: fixed support for percent selector for metadata profiles.
   * Breaking Changes: